requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
alembic
httpx
requests
PyGithub
tenacity
loguru
radon
pytest
pytest-asyncio
black
isort
langchain
langgraph
langchain-community
langchain-groq